upgrade and organized deps, but now having some problems with jest tests
