+++ o1js/app/zknotary/package.json
@@ -36,16 +36,17 @@
     "zustand": "^4.5.2"
   },
   "devDependencies": {
-    "@types/node": "^20",
+    "@types/node": "^20.10.0",
     "@types/react": "^18",
     "@types/react-dom": "^18",
-    "eslint": "^8",
+    "eslint": "^8.7.0",
     "eslint-config-next": "14.2.2",
+    "o1js": "0.15.3",
     "postcss": "^8",
     "tailwindcss": "^3.4.1",
-    "typescript": "^5"
+    "typescript": "^5.3.2"
   },
   "peerDependencies": {
-    "o1js": "0.18.*"
+    "o1js": "0.15.3"
   }
 }
