+++ o1js/contracts/contracts/.eslintrc.cjs
@@ -1,5 +1,4 @@
 module.exports = {
-  root: true,
   env: {
     browser: true,
     node: true,
