--- o1js/contracts/contracts/babel.config.cjs
@@ -1,6 +1,0 @@
-module.exports = {
-  presets: [
-    ['@babel/preset-env', { targets: { node: 'current' } }],
-    '@babel/preset-typescript',
-  ],
-};
