+++ o1js/contracts/contracts/tsup.config.ts
@@ -1,0 +1,9 @@
+import { defineConfig } from 'tsup';
+
+export default defineConfig((opts) => ({
+  entry: ['./src/**/*'],
+  dts: true,
+  format: ['esm'],
+  clean: !opts.watch,
+  outDir: 'build/src',
+}));
