+++ o1js/package.json
@@ -4,18 +4,12 @@
   "version": "1.0.0",
   "description": "",
   "scripts": {
-    "dev": "turbo dev --parallel",
+    "dev": "turbo dev --filter apps",
     "start": "turbo start",
     "clean:workspaces": "turbo clean",
     "typecheck": "turbo typecheck",
-    "lint": "turbo lint"
-  },
-  "devDependencies": {
-    "@turbo/gen": "1.13.2",
-    "@types/node": "^20.10.0",
-    "@types/json-stable-stringify": "^1.0.36",
-    "turbo": "1.13.2",
-    "typescript": "^5.3.2"
+    "lint": "turbo lint",
+    "postinstall": "manypkg check"
   },
   "engines": {
     "pnpm": "8.15.1",
@@ -23,5 +17,13 @@
   },
   "keywords": [],
   "author": "",
-  "license": "ISC"
+  "license": "Apache-2.0",
+  "dependencies": {
+    "@manypkg/cli": "^0.21.4",
+    "@turbo/gen": "1.13.2",
+    "@types/json-stable-stringify": "^1.0.36",
+    "@types/node": "^20.10.0",
+    "turbo": "1.13.2",
+    "typescript": "^5.3.2"
+  }
 }
